PolyLine2DAdjusterにRemoveModeを追加
