--- Assets/Scenes/PolygonAdjuster.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: e1f6ad6d950a44cbcb487874790cc7e0
-folderAsset: yes
-timeCreated: 1475072267
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
