--- Assets/Scenes/PolygonAdjuster/Chap1_PolygonAdjuster.unity.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: a63f9bdbcb15741f99b0c5d2fa2a73d9
-timeCreated: 1475072308
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
