--- Assets/Scenes/PolygonEditor.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 73a009c35062f4add9f9eaf363c3cdeb
-folderAsset: yes
-timeCreated: 1475073884
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
