--- Assets/Scenes/PolygonEditor/Chap1_PolygonEditor.unity.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 0a363410b4b3d4899835148d33b80c45
-timeCreated: 1475073902
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
