--- Assets/Scenes/PolygonMaker/Chap1_PolygonMaker.unity.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 78643cbdb928046f984350e5a21cd832
-timeCreated: 1475072179
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
