--- Assets/Scripts/LineEditor/Snap/LineSegSnap.cs.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-guid: 56c251d4f5b474c94b043ddab49c8d6b
-timeCreated: 1474710140
-licenseType: Free
-MonoImporter:
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
