--- Assets/Scripts/PolygonEditor.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 1590135db8d1444e9887524dafd16e0c
-folderAsset: yes
-timeCreated: 1475082783
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
