--- Assets/Scripts/PolygonMaker.meta
@@ -1,9 +1,0 @@
-fileFormatVersion: 2
-guid: 8a0d92f9148e74b6088489870dfe6172
-folderAsset: yes
-timeCreated: 1474727478
-licenseType: Free
-DefaultImporter:
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
