+++ Assets/Seiro/Scripts/Geometric/Polygon/Concave/ConcavePolygonEditor.cs
@@ -1,0 +1,15 @@
+﻿using UnityEngine;
+using Seiro.Scripts.Graphics.PolyLine2D;
+
+namespace Seiro.Scripts.Geometric.Polygon.Concave {
+
+	/// <summary>
+	/// 凹多角形の編集
+	/// </summary>
+	public class ConcavePolygonEditor : MonoBehaviour {
+
+		[Header("PolyLineEditor")]
+		public PolyLine2DEditor lineEditor;
+
+	}
+}
