+++ Assets/Seiro/Scripts/Graphics/PolyLine2D/Snap.meta
@@ -1,7 +1,7 @@
 fileFormatVersion: 2
-guid: 26f4a34a0044a46dc8ffad48f0a25154
+guid: 81b0f204761264ceb839702a166e24ef
 folderAsset: yes
-timeCreated: 1475072179
+timeCreated: 1475379002
 licenseType: Free
 DefaultImporter:
   userData:
