+++ Assets/Seiro/Scripts/Graphics/PolyLine2D/Snap/BaseSnap.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: b1c4b2c9e8a2e48799cf21cdb6364591
-timeCreated: 1474710140
+guid: e57d7e31ef2994853a79a5a536c413d2
+timeCreated: 1475379002
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
