+++ Assets/Seiro/Scripts/Graphics/PolyLine2D/Snap/PointSnap.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 6a040bac521314a838038d0ea1dd8a8c
-timeCreated: 1475072398
+guid: f80686b706daf484b94971b091a263ca
+timeCreated: 1475379002
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
