+++ Assets/Seiro/Scripts/Graphics/PolyLine2D/Snap/RadialSnap.cs.meta
@@ -1,6 +1,6 @@
 fileFormatVersion: 2
-guid: 59c69033d28834a03b779ee48daf8d5f
-timeCreated: 1474710140
+guid: fd9a41c4d7b1441a4bac48a995abf5a0
+timeCreated: 1475379002
 licenseType: Free
 MonoImporter:
   serializedVersion: 2
